Split up popover in to smaller components
