+++ src/plugins/controls/public/options_list/components/options_list_popover_suggestions.tsx
@@ -1,0 +1,123 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import React, { useMemo } from 'react';
+
+import { EuiFilterSelectItem, EuiSpacer, EuiIcon } from '@elastic/eui';
+import { useReduxEmbeddableContext } from '@kbn/presentation-util-plugin/public';
+
+import { OptionsListReduxState } from '../types';
+import { OptionsListStrings } from './options_list_strings';
+import { optionsListReducers } from '../options_list_reducers';
+
+interface OptionsListPopoverSuggestionsProps {
+  showOnlySelected: boolean;
+}
+
+export const OptionsListPopoverSuggestions = ({
+  showOnlySelected,
+}: OptionsListPopoverSuggestionsProps) => {
+  // Redux embeddable container Context
+  const {
+    useEmbeddableDispatch,
+    useEmbeddableSelector: select,
+    actions: { replaceSelection, deselectOption, selectExists, selectOption },
+  } = useReduxEmbeddableContext<OptionsListReduxState, typeof optionsListReducers>();
+  const dispatch = useEmbeddableDispatch();
+
+  // Select current state from Redux using multiple selectors to avoid rerenders.
+  const invalidSelections = select((state) => state.componentState.invalidSelections);
+  const availableOptions = select((state) => state.componentState.availableOptions);
+
+  const selectedOptions = select((state) => state.explicitInput.selectedOptions);
+  const existsSelected = select((state) => state.explicitInput.existsSelected);
+  const singleSelect = select((state) => state.explicitInput.singleSelect);
+  const hideExists = select((state) => state.explicitInput.hideExists);
+  const exclude = select((state) => state.explicitInput.exclude);
+
+  const loading = select((state) => state.output.loading);
+
+  // track selectedOptions and invalidSelections in sets for more efficient lookup
+  const selectedOptionsSet = useMemo(() => new Set<string>(selectedOptions), [selectedOptions]);
+  const invalidSelectionsSet = useMemo(
+    () => new Set<string>(invalidSelections),
+    [invalidSelections]
+  );
+  const suggestions = showOnlySelected ? selectedOptions : availableOptions;
+
+  if (!loading && (!suggestions || suggestions.length === 0) && !existsSelected) {
+    return (
+      <div
+        className="euiFilterSelect__note"
+        data-test-subj={`optionsList-control-${
+          showOnlySelected ? 'selectionsEmptyMessage' : 'noSelectionsMessage'
+        }`}
+      >
+        <div className="euiFilterSelect__noteContent">
+          <EuiIcon type="minusInCircle" />
+          <EuiSpacer size="xs" />
+          <p>
+            {showOnlySelected
+              ? OptionsListStrings.popover.getSelectionsEmptyMessage()
+              : OptionsListStrings.popover.getEmptyMessage()}
+          </p>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <>
+      {!hideExists && !(showOnlySelected && !existsSelected) && (
+        <EuiFilterSelectItem
+          data-test-subj={`optionsList-control-selection-exists`}
+          checked={existsSelected ? 'on' : undefined}
+          key={'exists-option'}
+          onClick={() => {
+            dispatch(selectExists(!Boolean(existsSelected)));
+          }}
+        >
+          <span className="optionsList__existsFilter">
+            {exclude
+              ? OptionsListStrings.controlAndPopover.getNegateExists()
+              : OptionsListStrings.controlAndPopover.getExists()}
+          </span>
+        </EuiFilterSelectItem>
+      )}
+      {suggestions?.map((suggestion, index) => (
+        <EuiFilterSelectItem
+          data-test-subj={`optionsList-control-selection-${suggestion}`}
+          checked={selectedOptionsSet?.has(suggestion) ? 'on' : undefined}
+          key={index}
+          onClick={() => {
+            if (showOnlySelected) {
+              dispatch(deselectOption(suggestion));
+              return;
+            }
+            if (singleSelect) {
+              dispatch(replaceSelection(suggestion));
+              return;
+            }
+            if (selectedOptionsSet.has(suggestion)) {
+              dispatch(deselectOption(suggestion));
+              return;
+            }
+            dispatch(selectOption(suggestion));
+          }}
+          className={
+            showOnlySelected && invalidSelectionsSet.has(suggestion)
+              ? 'optionsList__selectionInvalid'
+              : undefined
+          }
+        >
+          {`${suggestion}`}
+        </EuiFilterSelectItem>
+      ))}
+    </>
+  );
+};
